Initial implementation of `distance`
This file contains **three** implementations of distance. One, I think
is my implementation. Another is the implementation provided in the "answers". The
third is an implementation suggested by Google AI.

I do not know which implementation is actually correct; I do not have
sufficient domain knowledge to "make the call."